+++ 05/linear_search_array.c
@@ -2,17 +2,17 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-// size_t型: 恐らく64bitのシステムだとlong unsigned int, 32bitのシステムだとunsigned int.
+// int型: 恐らく64bitのシステムだとlong unsigned int, 32bitのシステムだとunsigned int.
 
 /**
- * @brief 指定されたsize_tのポインタにある配列にfiesher-yates shuffle.
- * @param[in,out] array シャフルする配列であるsize_tのポインタ.
+ * @brief 指定されたintのポインタにある配列にfiesher-yates shuffle.
+ * @param[in,out] array シャフルする配列であるintのポインタ.
  * @param[in] array_size シャッフルする配列のサイズ
  */
-void fisher_yates_shuffle(size_t* array, size_t array_size) {
-  size_t i = array_size;
+void fisher_yates_shuffle(int* array, int array_size) {
+  int i = array_size;
   while (i > 1) {
-    size_t j = rand() % i;
+    int j = rand() % i;
     i--;
     int t = array[i];
     array[i] = array[j];
@@ -46,7 +46,7 @@ void fisher_yates_shuffle(size_t* array, size_t array_size) {
  * @brief レコード：keyとfieldをもった構造体
  */
 typedef struct record {
-  size_t key;                   /** size_t型のキー. key==-1or2^64-1を例外処理に用いている*/
+  int key;                      /** int型のキー. key==-1or2^64-1を例外処理に用いている*/
   char field[MAX_FIELD_MEMORY]; /** データを保持するchar型の配列. */
 } record;
 
@@ -54,8 +54,8 @@ typedef struct record {
  * @brief 表：recordの配列とそのメタデータを保持する構造体.
  */
 typedef struct table {
-  size_t records_length;                             /** 配列の長さ. */
-  size_t size;                                       /** 配列がとれる最大長. */
+  int records_length;                                /** 配列の長さ. */
+  int size;                                          /** 配列がとれる最大長. */
   record records[MAX_ARRAY_MEMORY / sizeof(record)]; /** record型で表された配列. */
 } table;
 
@@ -65,7 +65,7 @@ typedef struct table {
  * @param[in] field const char*,動的にするのであればchar*,サイズはMAX_FIELD_MEMORYで定義.
  * @return 初期化されたrecordのポインタ.
  */
-record* init_record(size_t key, const char* field) {
+record* init_record(int key, const char* field) {
   if (strlen(field) > MAX_FIELD_MEMORY + 1) {
     fprintf(stderr, "ERROR: \"field\" is too large.\n");
     exit(1);
@@ -104,9 +104,9 @@ void release_table(table* tab) {
  * @param[in] target 調べるキー.
  * @return targetがrecords内に存在するかを示すbool.
  */
-bool search_existence(table* tab, size_t target) {
+bool search_existence(table* tab, int target) {
   tab->records[tab->records_length].key = target;
-  size_t searching_pos = 0;
+  int searching_pos = 0;
   // while(searching_pos < tab->records_length && target != tab->records[searching_pos].key)
   while (target != tab->records[searching_pos].key) {
     searching_pos++;
@@ -120,9 +120,9 @@ bool search_existence(table* tab, size_t target) {
  * @param[in] target 調べるキー.
  * @return targetを指すキーか,無い場合は-1==2^64-1,恐らくシステムが32bitか64bitで変化(あくまで予想).
  */
-size_t search_index(table* tab, size_t target) {
+int search_index(table* tab, int target) {
   tab->records[tab->records_length].key = target;
-  size_t searching_pos = 0;
+  int searching_pos = 0;
   while (target != tab->records[searching_pos].key) {
     searching_pos++;
   }
@@ -150,8 +150,8 @@ void insert_tail(table* tab, record* rec) {
  * @param[in] tab 削除対象のrecordsをもつtableのポインタ.
  * @param[in] pos 削除するrecordsの位置.
  */
-void delete_record(table* tab, size_t pos) {
-  for (size_t i = pos; i < tab->records_length - 1; i++) {
+void delete_record(table* tab, int pos) {
+  for (int i = pos; i < tab->records_length - 1; i++) {
     tab->records[i] = tab->records[i + 1];
   }
   tab->records_length--;
@@ -163,7 +163,7 @@ void delete_record(table* tab, size_t pos) {
  */
 record* cli_record() {
   record* rec;
-  size_t key = -1;
+  int key = -1;
   char field[MAX_FIELD_MEMORY];
 
   printf("Type in a key >= 0 and a field. (example: \"10001 BBB\")\n");
@@ -212,7 +212,7 @@ void print_record(record* rec) {
 void print_table(table* tab) {
   printf("================================\n");
   printf("TABLE: [\n");
-  for (size_t i = 0; i < tab->records_length; i++) {
+  for (int i = 0; i < tab->records_length; i++) {
     print_record(&tab->records[i]);
   }
   printf("]\nTABLE LENGTH: %ld\n", tab->records_length);
@@ -224,21 +224,21 @@ void print_table(table* tab) {
  * @param[in] tab targetが存在するか調べるtable.
  * @param[in] target 調べるキー.
  */
-void print_search_existence(table* tab, size_t target) {
+void print_search_existence(table* tab, int target) {
   bool b = search_existence(tab, target);
   printf("\"%zu\" was %s\n", target, b ? "FOUND." : "NOT FOUND.");
 }
 
 int main() {
-  size_t keys[SAMPLE_RECORDS_SIZE];
-  for (size_t i = 0; i < SAMPLE_RECORDS_SIZE; i++) {
+  int keys[SAMPLE_RECORDS_SIZE];
+  for (int i = 0; i < SAMPLE_RECORDS_SIZE; i++) {
     keys[i] = i;
   }
-  fisher_yates_shuffle(keys, sizeof(keys) / sizeof(size_t));
+  fisher_yates_shuffle(keys, sizeof(keys) / sizeof(int));
 
   table* tab = init_table();
   record* rec = NULL;
-  for (size_t i = 0; i < SAMPLE_RECORDS_SIZE; i++) {
+  for (int i = 0; i < SAMPLE_RECORDS_SIZE; i++) {
     rec = init_record(keys[i], "AAA");
     insert_tail(tab, rec);
   }
