+++ 06/binary_search_array.c
@@ -2,7 +2,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-// size_t型: 恐らく64bitのシステムだとlong unsigned int, 32bitのシステムだとunsigned int.
 
 /**
  * @brief 文字列化.
@@ -30,7 +29,7 @@
  * @brief レコード：keyとfieldをもった構造体
  */
 typedef struct record {
-  size_t key;                   /** size_t型のキー. key==-1or2^64-1を例外処理に用いている*/
+  int key;                      /** int型のキー. key==-1or2^64-1を例外処理に用いている*/
   char field[MAX_FIELD_MEMORY]; /** データを保持するchar型の配列. */
 } record;
 
@@ -38,8 +37,8 @@ typedef struct record {
  * @brief 表：recordの配列とそのメタデータを保持する構造体.
  */
 typedef struct table {
-  size_t records_length;                             /** 配列の長さ. */
-  size_t size;                                       /** 配列がとれる最大長. */
+  int records_length;                                /** 配列の長さ. */
+  int size;                                          /** 配列がとれる最大長. */
   record records[MAX_ARRAY_MEMORY / sizeof(record)]; /** record型で表された配列. */
 } table;
 
@@ -49,7 +48,7 @@ typedef struct table {
  * @param[in] field const char*,動的にするのであればchar*,サイズはMAX_FIELD_MEMORYで定義.
  * @return 初期化されたrecordのポインタ.
  */
-record *init_record(size_t key, const char *field) {
+record *init_record(int key, const char *field) {
   if (strlen(field) > MAX_FIELD_MEMORY + 1) {
     fprintf(stderr, "ERROR: \"field\" is too large.\n");
     exit(1);
@@ -89,9 +88,9 @@ void release_table(table *tab) {
  * @param[out] found targetがtab内に存在するかの真理値.
  * @param[out] index targetが存在する場合キーのインデックスを指す,無い場合は-1==2^64-1.
  */
-void linear_search_existence_and_index(table *tab, size_t target, bool *found, size_t *index) {
+void linear_search_existence_and_index(table *tab, int target, bool *found, int *index) {
   tab->records[tab->records_length].key = target;
-  size_t searching_pos = 0;
+  int searching_pos = 0;
   while (target != tab->records[searching_pos].key) {
     searching_pos++;
   }
@@ -106,7 +105,7 @@ void linear_search_existence_and_index(table *tab, size_t target, bool *found, s
  * @param[out] found targetがtab内に存在するかの真理値.
  * @param[out] index targetが存在する場合キーのインデックスを指す,無い場合はtarget以上の最小キー.
  */
-void binary_search_existence_and_index(table *tab, size_t target, bool *found, size_t *index) {
+void binary_search_existence_and_index(table *tab, int target, bool *found, int *index) {
   long int lo = 0, hi = tab->records_length - 1, mid;
   while (lo <= hi) {
     // whileから出るワンステップ前になりえる状態はlo==hiまたはlo==hi+1(lo==mid+1で前者になる).
@@ -130,9 +129,9 @@ void binary_search_existence_and_index(table *tab, size_t target, bool *found, s
   // okは常に条件を満たすインデックス
   // ngは常に条件を満たさないインデックス
   /*
-  size_t ng = -1;
-  size_t ok = tab->records_length; // tab->records[tab->records_length - 1].key < targetのことを考えて
-  size_t mid;
+  int ng = -1;
+  int ok = tab->records_length; // tab->records[tab->records_length - 1].key < targetのことを考えて
+  int mid;
   while (ok - ng > 1)
   {
     mid = (ok + ng) / 2;
@@ -163,10 +162,10 @@ void insert_record(table *tab, record *rec) {
   }
 
   bool dummy;
-  size_t inserting_pos;
+  int inserting_pos;
   binary_search_existence_and_index(tab, rec->key, &dummy, &inserting_pos);
 
-  for (size_t i = tab->records_length; i > inserting_pos; i--) {
+  for (int i = tab->records_length; i > inserting_pos; i--) {
     tab->records[i] = tab->records[i - 1];
   }
 
@@ -179,12 +178,12 @@ void insert_record(table *tab, record *rec) {
  * @param[in] tab 削除対象のrecordsをもつtableのポインタ.
  * @param[in] pos 削除するrecordsの位置.
  */
-void delete_record(table *tab, size_t pos) {
-  if (pos == (size_t)-1) {
+void delete_record(table *tab, int pos) {
+  if (pos == (int)-1) {
     printf("Valid pos was not set to delte. pos=\"%zu\".\n", pos);
     return;
   }
-  for (size_t i = pos; i < tab->records_length - 1; i++) {
+  for (int i = pos; i < tab->records_length - 1; i++) {
     tab->records[i] = tab->records[i + 1];
   }
   tab->records_length--;
@@ -196,7 +195,7 @@ void delete_record(table *tab, size_t pos) {
  */
 record *cli_record() {
   record *rec;
-  size_t key = -1;
+  int key = -1;
   char field[MAX_FIELD_MEMORY];
 
   printf("Type in a key >= 0 and a field. (example: \"10001 BBB\")\n");
@@ -222,7 +221,7 @@ void cli_insert(table *tab) {
   bool does_scanned_rec_exist = false;
   while (true) {
     scanned_rec = cli_record();
-    size_t dummy;
+    int dummy;
     binary_search_existence_and_index(tab, scanned_rec->key, &does_scanned_rec_exist, &dummy);
     if (does_scanned_rec_exist) {
       printf("The key is already used.\n");
@@ -248,7 +247,7 @@ void print_record(record *rec) {
 void print_table(table *tab) {
   printf("================================\n");
   printf("TABLE: [\n");
-  for (size_t i = 0; i < tab->records_length; i++) {
+  for (int i = 0; i < tab->records_length; i++) {
     print_record(&tab->records[i]);
   }
   printf("]\nTABLE LENGTH: %ld\n", tab->records_length);
@@ -260,9 +259,9 @@ void print_table(table *tab) {
  * @param[in] tab targetが存在するか調べるtable.
  * @param[in] target 調べるキー.
  */
-void print_search_existence(table *tab, size_t target) {
+void print_search_existence(table *tab, int target) {
   bool b = false;
-  size_t dummy;
+  int dummy;
   binary_search_existence_and_index(tab, target, &b, &dummy);
   printf("\"%zu\" was %s\n", target, b ? "FOUND." : "NOT FOUND.");
 }
@@ -270,7 +269,7 @@ void print_search_existence(table *tab, size_t target) {
 int main() {
   table *tab = init_table();
   record *rec = NULL;
-  for (size_t i = 0; i < SAMPLE_RECORDS_SIZE; i++) {
+  for (int i = 0; i < SAMPLE_RECORDS_SIZE; i++) {
     rec = init_record(i, "AAA");
     insert_record(tab, rec);
   }
@@ -281,7 +280,7 @@ int main() {
   // print_table(tab);
 
   bool dummy;
-  size_t target_pos;
+  int target_pos;
   binary_search_existence_and_index(tab, 5, &dummy, &target_pos);
   delete_record(tab, target_pos);
