+++ 06/binary_tree.c
@@ -30,7 +30,7 @@ typedef node *tree;
  * @brief レコード：keyとfieldをもった構造体
  */
 typedef struct record {
-  size_t key;                   /** size_t型のキー. key==-1or2^64-1を例外処理に用いている*/
+  int key;                      /** int型のキー. key==-1or2^64-1を例外処理に用いている*/
   char field[MAX_FIELD_MEMORY]; /** データを保持するchar型の配列. */
 } record;
 
@@ -46,7 +46,7 @@ struct node {
  * @param[in] field const char*,動的にするのであればchar*,サイズはMAX_FIELD_MEMORYで定義.
  * @return 初期化されたrecordのポインタ.
  */
-record *init_record(size_t key, const char *field) {
+record *init_record(int key, const char *field) {
   if (strlen(field) > MAX_FIELD_MEMORY + 1) {
     fprintf(stderr, "ERROR: \"field\" is too large.\n");
     exit(1);
@@ -168,7 +168,7 @@ tree extract_max_node(tree *rt) {
  * @param[in] rt 削除対象をもつtreeのポインタ.
  * @param[in] target 調べるrec->key.
  */
-void delete_node(tree *rt, size_t target) {
+void delete_node(tree *rt, int target) {
   tree *p = rt;
   while (*p != NULL) {
     if (target != (*p)->rec->key) {
@@ -204,7 +204,7 @@ void delete_node(tree *rt, size_t target) {
  * @param[in] target 調べるキー.
  * @return 存在する場合:対応するnode*,存在しない場合:NULL.
  */
-node *search_node(tree rt, size_t target) {
+node *search_node(tree rt, int target) {
   tree p = rt;
   while (p != NULL) {
     if (target == p->rec->key) {
@@ -238,7 +238,7 @@ void print_node(node *n) {
  * @param[in] rt targetが存在するか調べるtree.
  * @param[in] target 調べるキー.
  */
-void print_search_node(tree rt, size_t target) {
+void print_search_node(tree rt, int target) {
   tree search_result_node = search_node(rt, target);
   printf("\"%zu\" was %s\n", target, search_result_node != NULL ? "FOUND." : "NOT FOUND.");
 }
@@ -248,11 +248,11 @@ void print_search_node(tree rt, size_t target) {
  * @param[in] p 操作する対象のtree.
  * @param[in] depth 再帰呼び出しの深さ.
  */
-void print_tree_recursion(tree p, size_t depth) {
+void print_tree_recursion(tree p, int depth) {
   if (p != NULL) {
     print_tree_recursion(p->right, depth + 1);
 
-    for (size_t i = 0; i < depth; i++) {
+    for (int i = 0; i < depth; i++) {
       printf(" ");
     }
     printf("+-%3zu, \"%s\"\n", p->rec->key, p->rec->field);
@@ -265,7 +265,7 @@ void print_tree_recursion(tree p, size_t depth) {
  * @param[in] rt 操作する対象のtree.
  */
 void print_tree(tree rt) {
-  size_t depth = 0;
+  int depth = 0;
   printf("VISUALISING TREE\n");
   printf("================================\n");
   print_tree_recursion(rt, depth);
@@ -278,7 +278,7 @@ void print_tree(tree rt) {
  */
 record *cli_record() {
   record *rec;
-  size_t key = -1;
+  int key = -1;
   char field[MAX_FIELD_MEMORY];
 
   printf("Type in a key >= 0 and a field. (example: \"10001 BBB\")\n");
@@ -314,9 +314,9 @@ void cli_insert(tree *rt) {
 }
 
 int main() {
-  size_t tmp[] = {44, 55, 12, 42, 14, 18, 06, 67};
+  int tmp[] = {44, 55, 12, 42, 14, 18, 06, 67};
   tree rt = NULL;
-  for (size_t i = 0; i < sizeof(tmp) / sizeof(size_t); i++) {
+  for (int i = 0; i < sizeof(tmp) / sizeof(int); i++) {
     insert_node(&rt, init_node(init_record(tmp[i], "AAAA")));
   }
   print_tree(rt);
