+++ 09/sort_algos.h
@@ -6,18 +6,18 @@
 typedef struct record record;
 typedef struct sequence sequence;
 
-record *init_record(size_t key, char *field);
+record *init_record(int key, char *field);
 sequence *init_sequence();
 void release_record(record **rec);
 void release_sequence(sequence **seq);
 
 bool is_record_a_larger_than_b(record *a, record *b);
-bool is_a_larger_than_b(record **records, size_t a, size_t b);
-void compare_and_swap(record **records, size_t a, size_t b);
+bool is_a_larger_than_b(record **records, int a, int b);
+void compare_and_swap(record **records, int a, int b);
 
 void insert_tail(sequence *seq, record *rec);
 
-void binary_search_existence_and_index(sequence *seq, size_t target, bool *found, size_t *index);
+void binary_search_existence_and_index(sequence *seq, int target, bool *found, int *index);
 
 void simple_sort(sequence *seq);
 void bubble_sort(sequence *seq);
@@ -26,11 +26,11 @@ void insertion_sort(sequence *seq);
 
 void print_record(record *rec);
 void print_sequence(sequence *seq);
-void print_search_existence_and_index(sequence *seq, size_t target);
+void print_search_existence_and_index(sequence *seq, int target);
 
 record *cli_record();
 void cli_insert(sequence *seq);
 void run_sort_algorithm(sequence *seq, void (*sort_algorithm)(sequence *));
 
-void fisher_yates_shuffle(size_t *array, size_t array_size);
+void fisher_yates_shuffle(int *array, int array_size);
 #endif
